Conflicts:
	app/controllers/ResourcesController.php

+++ app/controllers/CategoriesController.php
@@ -9,10 +9,15 @@ class CategoriesController extends BaseController {
 	 */
 	public function index()
 	{
-		//
-		$categories = Category::all();
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			//
+			$categories = Category::all();
 
-		return View::make('admin.categories.index')->with('categories',$categories);
+			return View::make('admin.categories.index')->with('categories',$categories);
+		}else{
+			return Redirect::to('login');
+		}
 	}
 
 
@@ -23,8 +28,13 @@ class CategoriesController extends BaseController {
 	 */
 	public function create()
 	{
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
 		//
-		return View::make('admin.categories.create');
+			return View::make('admin.categories.create');
+		}else{
+			return Redirect::to('login');
+		}
 	}
 
 
@@ -35,31 +45,36 @@ class CategoriesController extends BaseController {
 	 */
 	public function store()
 	{
-		//
-		//
-		$rules = array(
-            'name'      => 'required',
-        );        
-        $messages = [
-        	'required' 	=> 'Este campo es obligatorio',
-        	'email'	   	=> 'Correo no valido',
-        	'min'	   	=> 'La matrícula no es valida',
-        	'size'	=> 'La matrícula no es valida'
-       	];
-		$validator = Validator::make(Input::all(), $rules, $messages);
-		
-		if($validator->fails()){
-			return Redirect::to('categories/create')
-			->withErrors($validator->messages())
-			->withInput();
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			//
+			//
+			$rules = array(
+	            'name'      => 'required',
+	        );        
+	        $messages = [
+	        	'required' 	=> 'Este campo es obligatorio',
+	        	'email'	   	=> 'Correo no valido',
+	        	'min'	   	=> 'La matrícula no es valida',
+	        	'size'	=> 'La matrícula no es valida'
+	       	];
+			$validator = Validator::make(Input::all(), $rules, $messages);
+			
+			if($validator->fails()){
+				return Redirect::to('categories/create')
+				->withErrors($validator->messages())
+				->withInput();
+			}else{
+				$category = new Category;
+				$category->name		=	Input::get('name');
+				$category->description	=	Input::get('description');			
+				$category->save(); 
+				Session::flash('message', 'Successfully created category!');
+				return Redirect::to('categories');
+			}	
 		}else{
-			$category = new Category;
-			$category->name		=	Input::get('name');
-			$category->description	=	Input::get('description');			
-			$category->save(); 
-			Session::flash('message', 'Successfully created category!');
-			return Redirect::to('categories');
-		}	
+			return Redirect::to('login');
+		}
 	}
 
 
@@ -71,9 +86,14 @@ class CategoriesController extends BaseController {
 	 */
 	public function show($id)
 	{
-		//
-		$category = Category::find($id);
-		return View::make('admin.categories.show')->with('category',$category);
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			//
+			$category = Category::find($id);
+			return View::make('admin.categories.show')->with('category',$category);
+		}else{
+			return Redirect::to('login');
+		}
 	}
 
 
@@ -85,9 +105,14 @@ class CategoriesController extends BaseController {
 	 */
 	public function edit($id)
 	{
-		//
-		$category = Category::find($id);
-		return View::make('admin.categories.edit')->with('category',$category);
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			//
+			$category = Category::find($id);
+			return View::make('admin.categories.edit')->with('category',$category);
+		}else{
+			return Redirect::to('login');
+		}
 	}
 
 
@@ -99,38 +124,48 @@ class CategoriesController extends BaseController {
 	 */
 	public function update($id)
 	{
-		//
-		$rules = array(
-            'name'      => 'required',
-        );        
-        $messages = [
-        	'required' 	=> 'Este campo es obligatorio',
-        	'email'	   	=> 'Correo no valido',
-        	'min'	   	=> 'La matrícula no es valida',
-        	'size'	=> 'La matrícula no es valida'
-       	];
-		$validator = Validator::make(Input::all(), $rules, $messages);
-		
-		if($validator->fails()){
-			return Redirect::to('categories/' . $id . '/edit')
-			->withErrors($validator->messages())
-			->withInput();
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			//
+			$rules = array(
+	            'name'      => 'required',
+	        );        
+	        $messages = [
+	        	'required' 	=> 'Este campo es obligatorio',
+	        	'email'	   	=> 'Correo no valido',
+	        	'min'	   	=> 'La matrícula no es valida',
+	        	'size'	=> 'La matrícula no es valida'
+	       	];
+			$validator = Validator::make(Input::all(), $rules, $messages);
+			
+			if($validator->fails()){
+				return Redirect::to('categories/' . $id . '/edit')
+				->withErrors($validator->messages())
+				->withInput();
+			}else{
+				$category = Category::find($id);
+				$category->name		=	Input::get('name');
+				$category->description	=	Input::get('description');			
+				$category->save(); 
+				Session::flash('message', 'Successfully updated category!');
+				return Redirect::to('categories');
+			}		
 		}else{
-			$category = Category::find($id);
-			$category->name		=	Input::get('name');
-			$category->description	=	Input::get('description');			
-			$category->save(); 
-			Session::flash('message', 'Successfully updated category!');
-			return Redirect::to('categories');
-		}		
+			return Redirect::to('login');
+		}
 	}
 
 	public function destroy($id)
 	{
-		$category = Category::find($id);
-		$category->delete();
-		Session::flash('message', 'Successfully deleted category!');
-		return Redirect::to('categories');
+		if (Session::get('school_id') && Session::get('role')==1)
+		{
+			$category = Category::find($id);
+			$category->delete();
+			Session::flash('message', 'Successfully deleted category!');
+			return Redirect::to('categories');
+		}else{
+			return Redirect::to('login');
+		}
 	}
